+++ www/src/components/thumbprint-tokens/tab-nav/index.jsx
@@ -5,6 +5,7 @@ const TokensTabNav = () => (
     <TabNav>
         <TabNavItem to="/tokens/scss/">SCSS</TabNavItem>
         <TabNavItem to="/tokens/javascript/">JavaScript</TabNavItem>
+        <TabNavItem to="/tokens/ios/">iOS</TabNavItem>
     </TabNav>
 );