+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,11 +2,18 @@
 
 ## Unreleased
 
+### Fixed
+
+-   [Patch] Fix bug in `use-focus-trap` that caused it to refocus every render. (#366)
+
+## 6.4.0 - 2019-07-09
+
 ### Changed
 
 -   [Major] Rewrite `DatePicker` to be a fully controlled component. (#12)
 -   [Patch] Update `ConditionalPortal` to explicitly allow and handle null contents. (#357)
 -   [Patch] Update description of Avatar prop.
+-   [Minor] Remove fade-in from `Image`. (#368)
 
 ## 6.3.2 - 2019-07-01